src/test/java/minimumIndexSumOfTwoLists/SolutionTest.java
package minimumIndexSumOfTwoLists;

import MinimumIndexSumOfTwoLists.*;

import org.junit.Assert;
import org.junit.Test;

public class SolutionTest {
    @Test
    public void test1() {
        Solution s = new Solution();
        Assert.assertArrayEquals(new String[]{"Shogun"}, s.findRestaurant(new String[]{"Shogun","Tapioca Express","Burger King","KFC"},
                new String[]{"Piatti","The Grill at Torrey Pines","Hungry Hunter Steakhouse","Shogun"}));
    }

    @Test
    public void test2() {
        Solution s = new Solution();
        Assert.assertArrayEquals(new String[]{"Shogun"}, s.findRestaurant(new String[]{"Shogun","Tapioca Express","Burger King","KFC"},
                new String[]{"KFC","Shogun","Burger King"}));
    }

    @Test
    public void test3() {
        Solution s = new Solution();
        Assert.assertArrayEquals(new String[]{"happy", "sad"}, s.findRestaurant(new String[]{"happy","sad","good"},
                new String[]{"sad","happy","good"}));
    }
}
